Kernel: working vmm
The VMM, at least the mapping part of it, is in a function state. As
well as the now, fairly simple, invlpg. Made numerous silly mistakes
because of my macros and my code not obeying the flags structure. There
was also an issue where the PMM would some reason start allocating in
the first SLAB list, this was fixed by making the SLAB allocator
allocate memory for its lists using the PMM's wrapper.

+++ kernel/src/c/include/mm/pmm.h
@@ -32,9 +32,9 @@
 #include <global.h>
 
 void *Arc_AllocPMM();
-void *Arc_ContiguousAllocPMM(int objects);
+void *Arc_ContiguousAllocPMM(size_t objects);
 void *Arc_FreePMM(void *address);
-void *Arc_ContiguousFreePMM(void *address, int objects);
+void *Arc_ContiguousFreePMM(void *address, size_t objects);
 void Arc_InitPMM(struct multiboot_tag_mmap *mmap, uint32_t boot_state);
 
 #endif
